requirements.txt
python-dotenv==1.0.1
google-generativeai==0.8.3
vertexai==1.71.1
langchain-google-genai==2.0.8
pillow==11.1.0